Fixed bug

+++ data_processing/remove_drift.py
@@ -82,7 +82,7 @@ def time_adjustment(first_data_cube, coefficient_cube):
     branch_time_value = first_data_cube.attributes['branch_time'] #FIXME: Add half a time step?
     branch_time_unit = coefficient_cube.attributes['time_unit']
     branch_time_calendar = coefficient_cube.attributes['time_calendar']
-    data_time_coord = first_data_cube.coord('time').astype(numpy.float32)
+    data_time_coord = first_data_cube.coord('time')
 
     new_unit = cf_units.Unit(branch_time_unit, calendar=branch_time_calendar)  
     data_time_coord.convert_units(new_unit)
